requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
python-dotenv>=1.0.0

# Document Processing
pypdf>=3.17.0
pymupdf>=1.23.0

# Text Processing
langchain-text-splitters>=0.0.1

# Embeddings & Vector Store
sentence-transformers>=2.2.2
chromadb>=0.4.22

# Retrieval & Re-ranking
rank-bm25>=0.2.2

# ML & Data Processing
numpy>=1.24.0
scikit-learn>=1.3.0

# LLM Providers
groq>=0.4.0

# Optional: Better embeddings (uncomment if needed)
# openai>=1.0.0
# cohere>=4.0.0

# Web UI
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
